updated Ex004 changed float to double

+++ Ex004_div/Program.cs
@@ -1,4 +1,4 @@
-﻿float numberA = 12;
-float numberB = 5;
-float result = numberA / numberB;
+﻿double numberA = 12;
+double numberB = 5;
+double result = numberA / numberB;
 Console.WriteLine(result);
